+++ src/array/chunk_grid/regular.rs
@@ -222,6 +222,8 @@ impl ChunkGridTraits for RegularChunkGrid {
 
 #[cfg(test)]
 mod tests {
+    use crate::array_subset::ArraySubset;
+
     use super::*;
 
     #[test]
@@ -288,6 +290,36 @@ mod tests {
                 .unwrap(),
             Some(vec![0, 1, 2])
         );
+
+        assert_eq!(
+            chunk_grid
+                .chunks_subset(
+                    &ArraySubset::new_with_ranges(&[1..3, 1..2, 5..8]),
+                    &array_shape
+                )
+                .unwrap(),
+            Some(ArraySubset::new_with_ranges(&[1..3, 2..4, 15..24]))
+        );
+
+        assert!(chunk_grid
+            .chunks_subset(&ArraySubset::new_with_ranges(&[1..3]), &array_shape)
+            .is_err());
+
+        assert!(chunk_grid
+            .chunks_subset(
+                &ArraySubset::new_with_ranges(&[1..3, 1..2, 5..8]),
+                &vec![0; 1]
+            )
+            .is_err());
+
+        assert!(chunk_grid
+            .chunks_subset(
+                &ArraySubset::new_with_ranges(&[0..0, 0..0, 0..0]),
+                &array_shape
+            )
+            .unwrap()
+            .unwrap()
+            .is_empty());
     }
 
     #[test]